Modified display, added animations
The old way was garbage, a truly terrible implementation. It relied on Bootstrap's rows and made it just sorta bad. Plus, it was poorly done.

In my new implementation, each tile just gets offset and that's all good. Plus, since each tile is just offset instead of relying on the width of the tiles in that row to maintain cohesion, now animations are possible. I implemented a basic test one.